backend/requirements.txt
fastapi==0.120.2
uvicorn==0.38.0
pydantic==2.12.3
python-dotenv==1.2.1
google-generativeai==0.8.1
selenium>=4.0.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
zemberek-python==0.2.3
slowapi>=0.1.9
requests>=2.28.0
sentry-sdk>=1.40.0
fastembed>=0.2.0
tenacity>=8.0.0
redis>=5.0.0  # Opsiyonel: REDIS_URL env var tanımlandığında kullanılır
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.23
